+++ app/src/main/java/com/example/shimataro/eho/Eho.java
@@ -9,7 +9,7 @@ public class Eho {
     }
 
     private static final EHO   m_eho_symbols[] = {EHO.WSW, EHO.SSE, EHO.NNW, EHO.SSE, EHO.ENE};
-    private static final float m_eho_orientations[] = {(float)247.5, (float)157.5, (float)337.5, (float)157.5, (float)67.5};
+    private static final float m_eho_orientations[] = {247.5f, 157.5f, 337.5f, 157.5f, 67.5f};
     private int m_year = 0;
 
     public void setYear(final int year) {
